services/scrapers/events/filter/requirements.txt
python-dotenv
requests
pandas
psycopg2-binary
google-genai
google-cloud-storage
supabase